changed int type to float for decimal numbers

+++ C/product_of_a_list.c
@@ -6,11 +6,11 @@
 int main()
 {
     int lst[]={10, 5, 16, 23, 54, 6};
-    int product = 1,i;
+    float product = 1,i;
     for(i=0;i<6;i++)
     {
         product=product*lst[i];
     }
-    printf("%d",product); // Output: 5961600  
+    printf("%f",product); // Output: 5961600  
     return 0;
 }
